Fire extensions (#402)
* Add extensions directory and tests

This prepares the way for exposing frequently used schema extensions to
users in a structured way. The schema extensions are deemed not wide
ranging and uniform enough to include in the main fire schemas.

* Refactor tests to use parametrization

Removes `for schema in SCHEMAS` loops from tests, which makes it much
clearer which schema is failing when the tests fail.

* feat(extensions): add US specific extensions

* style(tests): lint the tests

* feat(extensions): update extension enum lists

* Merge duplicated extension properties

---------

Co-authored-by: oliver <oliver@suade.org>
Co-authored-by: albie <albie@suade.org>
Co-authored-by: murat <murat@suade.org>